+++ packages/base/src/features/story/presentation/index.ts
@@ -1,5 +1,13 @@
 export type { StoryPresentationMode } from './types';
-export type { IStoryPresentationRootProps, IStoryStageProps } from './types';
+export type {
+  IStoryPresentationRootProps,
+  IStoryStageProps,
+} from './types';
+export type {
+  IStoryPresentationChromeProps,
+  IStoryPresentationDesktopChromeProps,
+  IStoryPresentationMobileChromeProps,
+} from './sharedChromeProps';
 
 export {
   getStoryPresentationMode,
@@ -8,3 +16,10 @@ export {
 
 export { StoryStage } from './StoryStage';
 export { StoryPresentationRoot } from './StoryPresentationRoot';
+export {
+  StoryPresentationDesktopChrome,
+  StoryPresentationMobileChrome,
+} from './StoryPresentationChrome';
+
+export * from './modes/column';
+export * from './modes/verticalScroll';
