+++ packages/base/src/features/story/presentation/modes/verticalScroll/VerticalScrollPresentationMobile.tsx
@@ -1,34 +1,22 @@
-import { IJGISStoryMap, IJupyterGISModel } from '@jupytergis/schema';
 import React, { useLayoutEffect, useMemo, useRef } from 'react';
 
 import { ListStoryVirtualScrollTrack } from '@/src/features/story/components/ListStoryVirtualScrollTrack';
 import { useListStoryScrollTrackContext } from '@/src/features/story/context/ListStoryScrollTrackContext';
 import { useListStoryScroll } from '@/src/features/story/hooks/useListStoryScroll';
-import type { IListStorySegmentTransition } from '@/src/features/story/types/types';
+import type { IStoryPresentationMobileChromeProps } from '@/src/features/story/presentation/sharedChromeProps';
 import { getSpectaPresentationStyle } from '@/src/features/story/utils/spectaPresentation';
 import { buildStorySegmentViewItems } from '@/src/features/story/utils/storySegmentViewItems';
-import { STORY_TYPE } from '@/src/types';
-
-export interface ISpectaMobileListModeContentProps {
-  model: IJupyterGISModel;
-  storyData: IJGISStoryMap | null;
-  currentIndex: number;
-  setIndex: (index: number) => void;
-  onSegmentTransitionChange?: (
-    payload: IListStorySegmentTransition | null,
-  ) => void;
-}
 
 /**
- * Mobile list story: full-viewport touch scroll on the virtual track drives
+ * Vertical-scroll chrome on mobile: full-viewport touch scroll on the virtual track.
  */
-export function SpectaMobileListModeContent({
+export function VerticalScrollPresentationMobile({
   model,
   storyData,
   currentIndex,
   setIndex,
   onSegmentTransitionChange,
-}: ISpectaMobileListModeContentProps): JSX.Element {
+}: IStoryPresentationMobileChromeProps): JSX.Element {
   const scrollContainerRef = useRef<HTMLDivElement>(null);
 
   const { scrollTrackLayout, bindScrollTrackElement } =
@@ -41,10 +29,6 @@ export function SpectaMobileListModeContent({
 
   const presentationStyle = getSpectaPresentationStyle(storyData);
 
-  const segmentTransitionSyncEnabled =
-    Boolean(onSegmentTransitionChange) &&
-    storyData?.storyType === STORY_TYPE.verticalScroll;
-
   useLayoutEffect(() => {
     bindScrollTrackElement(scrollContainerRef.current);
     return () => {
@@ -53,7 +37,7 @@ export function SpectaMobileListModeContent({
   }, [bindScrollTrackElement]);
 
   useListStoryScroll({
-    enabled: segmentTransitionSyncEnabled,
+    enabled: Boolean(onSegmentTransitionChange),
     scrollContainerRef,
     storyData,
     scrollTrackLayout,
